updated basic profile with basicSetUp in controller folder, this should be all of the profile related paths/routes done/refactored
